Added partially completed track switching
The variable that decides on with paths are in place is located on Obj_RailCart_Track_Switch. When set to false the original paths work correctly. When set to true the new paths not working correctly. When C2->A1 path ends the C2 button does not reset which blocks the cart from moving to another position.